requirements.txt
# MediMind AI - Dependencies

# FastAPI Backend
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
python-dotenv>=1.0.1

# LangChain & AI
langchain>=0.3.7
langchain-google-genai>=2.0.4
langchain-community>=0.3.7
langchain-chroma>=0.1.4
langgraph>=0.2.0

# Vector DB & Embeddings
chromadb>=0.5.15
sentence-transformers>=3.3.0

# PDF Processing
pypdf>=5.1.0

# Medical Image Analysis
open-clip-torch>=2.26.1
Pillow>=11.0.0

# Voice Processing
elevenlabs>=2.0.0

# Web Search
tavily-python>=0.5.0

# Database
sqlalchemy>=2.0.36
aiosqlite>=0.20.0

# Utilities
httpx>=0.27.2
pydantic>=2.10.0
pydantic-settings>=2.6.1
